[refactor] TestData -> SampleData 이동
