--- presentation/src/main/java/com/daejol/presentation/data/TestData.kt
@@ -1,46 +1,0 @@
-package com.daejol.presentation.data
-
-object TestData {
-    val animals = listOf(
-        Animal.Cat(
-            name = "Russian Blue",
-            image = "https://cdn2.thecatapi.com/images/cqg.jpg"
-        ),
-        Animal.Cat(
-            name = "Keykat",
-            image = "https://cdn2.thecatapi.com/images/ebv.jpg"
-        ),
-        Animal.Cat(
-            name = "Snow Cat",
-            image = "https://cdn.thedogapi.com/images/Hylo4Snaf.jpeg"
-        ),
-        Animal.Cat(
-            name = "Grasskitty",
-            image = "https://cdn2.thecatapi.com/images/xxBaNrfM0.jpg"
-        ),
-        Animal.Cat(
-            name = "Russian Blue",
-            image = "https://cdn2.thecatapi.com/images/cqg.jpg"
-        ),
-        Animal.Cat(
-            name = "Keykat",
-            image = "https://cdn2.thecatapi.com/images/ebv.jpg"
-        ),
-        Animal.Cat(
-            name = "Snow Cat",
-            image = "https://cdn.thedogapi.com/images/Hylo4Snaf.jpeg"
-        ),
-        Animal.Cat(
-            name = "Grasskitty",
-            image = "https://cdn2.thecatapi.com/images/xxBaNrfM0.jpg"
-        ),
-        Animal.Cat(
-            name = "Russian Blue",
-            image = "https://cdn2.thecatapi.com/images/cqg.jpg"
-        ),
-        Animal.Cat(
-            name = "Keykat",
-            image = "https://cdn2.thecatapi.com/images/ebv.jpg"
-        )
-    )
-}
